debate/requirements.txt
streamlit
requests
ollama
Flask
google-generativeai
flask-cors
